+++ src/js/Configuration.js
@@ -1,0 +1,7 @@
+const Configuration = {
+    ICON_ERROR_CLASS: "fa fa-exclamation-triangle",
+    ICON_VALID_CLASS: "fa fa-check",
+    COUNT_MESSAGE: (count) => `${count} character(s) remaining`
+};
+
+module.exports = Configuration;