+++ src/js/index.js
@@ -4,5 +4,6 @@ module.exports = {
     ValidationUtils         : require("./utils/ValidationUtils"),
     ValidationField         : require("./ValidationField"),
     ValidationSubmit        : require("./ValidationSubmit"),
-    ValidationTypes         : require("./ValidationTypes")
+    ValidationTypes         : require("./ValidationTypes"),
+    Config                  : require("./Configuration")
 };